+++ c/src/lib.rs
@@ -1,5 +1,8 @@
 mod cfile;
 pub mod compat;
 pub mod decode;
+pub mod encode;
 pub mod metadata;
+#[cfg(test)]
+mod test_utils;
 pub mod text_serialization;